LU-17398 build: detect mlnx-ofa_kernel-devel contents
Parse the configure_args for --with-o2ib and allow the
user specified path to override the mofed defaults.

Further when mlnx-ofa_kernel-devel contents are available silence
the BuildRequires: to allow for an mlnx source installation to
satisfy the lustre build requirements.

In addition move the mlnx specfic requirements to the mofed lnd
when '--with multiple_lnds' is enabled.

Fixes: 67cd54d05d ("LU-16967 build: Separate lnet LND rpm packaging")
Test-Parameters: trivial
Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: I30c6b3a196021634c621f6f6c556bf32f28faeed
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/53613
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Jian Yu <yujian@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>