LU-17000 utils: Remove check for errno != 0
In lustre_rename_fsname() after calls to system calls
like open/read/write/lseek there is a check if global
errno is not equal to zero. This is noop and not
required, as these system calls do reset errno to 0
on success.

The side effect of this check was that it appeared
that the errno could be 0 which would leave 'ret' as
negative. This would never happen, but was causing
Coverity to complain.

This patch fixes this coverity issue by removing
the comparison of errno != 0 which is not required.

Test-Parameters: trivial testlist=conf-sanity,sanityn
Fixes: d0c6e97fa53 ("LU-8900 snapshot: rename filesysetem fsname")
CoverityID: 397153 ("Argument cannot be negative")
Signed-off-by: Arshad Hussain <arshad.hussain@aeoncomputing.com>
Change-Id: I7257434e6546f56f30841f49a3bde35e80360bb8
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/53742
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Timothy Day <timday@amazon.com>
Reviewed-by: James Simmons <jsimmons@infradead.org>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>