LU-10391 lnet: Fix lnetctl peer set --all
When the --all flag is specified as part of the lnetctl peer set
command,  a primary NID argument is not sent to the peer doit command.
The "peer set" case uses NLM_F_REPLACE flag, so in this case the
primary NID argument is optional.

Added test cases to validate behavior of the lnetctl peer set
command.

Test-Parameters: trivial
Fixes: 8a0fdfa0b2 ("LU-10391 lnet: migrate peer NI control to Netlink")
Signed-off-by: Chris Horn <chris.horn@hpe.com>
Change-Id: I58de030b061280e837de27611bc701a3affab0f3
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/53791
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: James Simmons <jsimmons@infradead.org>
Reviewed-by: Frank Sehr <fsehr@whamcloud.com>
Reviewed-by: Cyril Bordage <cbordage@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>