LU-10366 test: re-enable sanity test 410 for Ubuntu
For older Ubuntu the pr_err() messages in the kinode module was not
making it to the dmesg ring buffer due to the default loglevel used
in their environment. Now that older verisons of Ubuntu are dropped
sanity 410 should pass. Thanks to the work of LU-17096 the kinode
modules is in its proper place so sanity 410 should pass.

This patch also changes test_410 and kinode to load the module
successfully and unload after the test is done. The Lustre spec
file is adjusted to accommodate this change.

Test-Parameters: trivial
Test-Parameters: testlist=sanity env=ONLY=410,ONLY_REPEAT=10
Test-Parameters: testlist=sanity env=ONLY=410,ONLY_REPEAT=10 clientdistro=ubuntu2204
Change-Id: Iac96efe64db721f9d7247a889f6e9bd4c7d45e2a
Signed-off-by: James Simmons <jsimmons@infradead.org>
Signed-off-by: Timothy Day <timday@amazon.com>
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/31921
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>