LU-15033 llite: Increase default RA sizes
The default max_readahead_mb is 1/32 of all cached pages,
which doesn't make much sense but isn't usually a problem
since most real nodes have very high RAM or it is tuned to
larger values.

It is reduced further for the per-file limit, which is also
reasonable.

However, on test VMs with smaller RAM sizes, this results
in hilariously tiny max_read_ahead_per_file_mb values, like
20.  This is small enough it causes extra misses because
two RPCs cannot be reliably sent.  This edge case isn't
important for performance, but it makes small scale testing
of readahead nearly impossible.

To avoid this, we add a minimum readahead requirement of
256 MiB, which is used unless it's > half of RAM.  This
should avoid this case on test VMs without changing the
behavior for real clients unless they are extremely small.

Signed-off-by: Patrick Farrell <pfarrell@whamcloud.com>
Change-Id: Ie8aab6b04ad520e4633d634d846e7ef23cc91ced
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/46475
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Sebastien Buisson <sbuisson@ddn.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>